requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1601924105527/work
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1606601395031/work
chardet==4.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///D:/bld/cryptography_1610338909973/work
cycler==0.10.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
docopt==0.6.2
flatbuffers==1.12
gast==0.3.3
google-api-core==1.25.1
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-cloud-automl==2.1.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.1
google-cloud-vision==2.1.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
html5lib @ file:///home/conda/feedstock_root/build_artifacts/html5lib_1592930327044/work
idna==2.10
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1612212935303/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
lastpass-python==0.3.1
libcst==0.3.16
Markdown==3.3.3
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mechanize==0.4.5
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbformat==5.1.2
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas @ file:///C:/ci/pandas_1611171485393/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pathlib==1.0.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pigar==0.10.0
Pillow @ file:///C:/ci/pillow_1609786840597/work
pipreqs==0.4.10
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pycryptodome==3.9.9
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent==0.17.3
PySocks @ file:///D:/bld/pysocks_1610291742246/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-slugify==4.0.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn @ file:///C:/ci/scikit-learn_1598377018496/work
scipy @ file:///C:/ci/scipy_1604596260408/work
selenium @ file:///D:/bld/selenium_1610664674276/work
sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
soupsieve==2.0.1
SQLAlchemy @ file:///C:/ci/sqlalchemy_1598374833867/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-serving-api==2.4.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.58.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1611695416663/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1610147566231/work
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
